requirements.txt
aiohttp==3.7.4.post0
APScheduler==3.7.0
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.2.0
cachetools==5.0.0
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
click==8.0.3
colorama==0.4.4
comtypes==1.1.10
cryptography==36.0.1
decorator==4.4.2
future==0.18.2
geojson==2.5.0
google-api-core==2.5.0
google-auth==2.6.0
google-cloud-dialogflow==2.11.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
huepy==1.2.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.4
imagesize==1.2.0
iniconfig==1.1.1
instabot==0.117.0
instagram==2.1.0
instagrapi==1.7.9
instauto==2.0.0
joblib==1.1.0
libretranslatepy==2.1.1
lxml==4.7.1
mock==4.0.3
moviepy==1.0.3
multidict==5.1.0
numpy==1.20.3
orjson==3.5.3
packaging==20.9
Pillow==8.2.0
pluggy==0.13.1
proglog==0.1.9
proto-plus==1.19.9
protobuf==3.19.4
py==1.10.0
pyanimation==0.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///C:/Users/Pavel%20Tkachev/AppData/Local/Programs/Python/Python39/Scripts/PyAudio-0.2.11-cp39-cp39-win_amd64.whl
pycparser==2.21
pycryptodomex==3.9.9
pydantic==1.8.2
pygame==2.0.1
Pyganim==0.9.2
PyJWT==2.3.0
pyowm==3.2.0
pyparsing==2.4.7
pypiwin32==223
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PySocks==1.7.1
pytest==6.2.4
pytest-asyncio==0.15.1
pytest-random-order==1.0.4
pyttsx3==2.90
pytz==2021.1
pywin32==303
regex==2022.1.18
requests==2.25.1
requests-toolbelt==0.9.1
responses==0.13.3
rsa==4.8
schedule==1.1.0
six==1.16.0
sounddevice==0.4.4
speech-recognition-fork==3.8.1.2021.6.14
toml==0.10.2
tqdm==4.61.0
translate==3.6.1
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.5
vosk==0.3.32
Wikipedia-API==0.5.4
yarl==1.6.3
